Use core objects for sn-client header
