+++ src/starknet/block.rs
@@ -1,5 +1,6 @@
 use serde::{Deserialize, Serialize};
 
+use super::serde_utils::{HexAsBytes, NonPrefixedHexAsBytes, PrefixedHexAsBytes};
 use super::{ContractAddress, StarkHash};
 
 // TODO(spapini): Verify the invariant that it is in range.
@@ -8,9 +9,25 @@ use super::{ContractAddress, StarkHash};
 )]
 pub struct BlockHash(pub StarkHash);
 #[derive(
-    Debug, Default, Copy, Clone, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord,
+    Debug, Copy, Clone, Default, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord,
+)]
+#[serde(
+    from = "NonPrefixedHexAsBytes<32_usize>",
+    into = "NonPrefixedHexAsBytes<32_usize>"
 )]
 pub struct GlobalRoot(pub StarkHash);
+// We don't use the regular StarkHash deserialization since the Starknet sequencer returns the
+// global root hash as a hex string without a "0x" prefix.
+impl From<NonPrefixedHexAsBytes<32_usize>> for GlobalRoot {
+    fn from(val: NonPrefixedHexAsBytes<32_usize>) -> Self {
+        GlobalRoot(StarkHash(val.0))
+    }
+}
+impl From<GlobalRoot> for NonPrefixedHexAsBytes<32_usize> {
+    fn from(val: GlobalRoot) -> Self {
+        HexAsBytes(val.0 .0)
+    }
+}
 #[derive(
     Debug, Default, Copy, Clone, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord,
 )]
@@ -32,9 +49,24 @@ impl BlockNumber {
 )]
 pub struct BlockTimestamp(pub u64);
 #[derive(
-    Debug, Default, Copy, Clone, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord,
+    Debug, Copy, Clone, Default, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord,
+)]
+#[serde(
+    from = "PrefixedHexAsBytes<16_usize>",
+    into = "PrefixedHexAsBytes<16_usize>"
 )]
 pub struct GasPrice(pub u128);
+impl From<PrefixedHexAsBytes<16_usize>> for GasPrice {
+    fn from(val: PrefixedHexAsBytes<16_usize>) -> Self {
+        GasPrice(u128::from_be_bytes(val.0))
+    }
+}
+impl From<GasPrice> for PrefixedHexAsBytes<16_usize> {
+    fn from(val: GasPrice) -> Self {
+        HexAsBytes(val.0.to_be_bytes())
+    }
+}
+
 #[derive(
     Debug, Default, Copy, Clone, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord,
 )]
