+++ src/starknet_client/objects/block.rs
@@ -2,11 +2,11 @@ use std::collections::HashMap;
 
 use serde::{Deserialize, Serialize};
 
-use crate::starknet;
-use crate::starknet::serde_utils::{HexAsBytes, NonPrefixedHexAsBytes, PrefixedHexAsBytes};
+use crate::starknet::{
+    BlockHash, BlockNumber, BlockTimestamp, ContractAddress, GasPrice, GlobalRoot, StarkHash,
+};
 
 use super::transaction::{ClassHash, Transaction, TransactionReceipt};
-use super::StarkHash;
 
 #[derive(Debug, Default, Deserialize, Serialize, Clone)]
 pub struct Block {
@@ -34,26 +34,6 @@ pub struct BlockStateUpdate {
     pub state_diff: StateDiff,
 }
 
-#[derive(
-    Debug, Copy, Clone, Default, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord,
-)]
-pub struct BlockHash(pub StarkHash);
-impl From<BlockHash> for starknet::BlockHash {
-    fn from(val: BlockHash) -> Self {
-        starknet::BlockHash(val.0.into())
-    }
-}
-
-#[derive(
-    Debug, Copy, Clone, Default, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord,
-)]
-pub struct BlockNumber(pub u64);
-impl From<BlockNumber> for starknet::BlockNumber {
-    fn from(val: BlockNumber) -> Self {
-        starknet::BlockNumber(val.0)
-    }
-}
-
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord)]
 pub enum BlockStatus {
     #[serde(rename(deserialize = "ABORTED", serialize = "ABORTED"))]
@@ -73,80 +53,10 @@ impl Default for BlockStatus {
     }
 }
 
-#[derive(
-    Debug, Copy, Clone, Default, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord,
-)]
-pub struct BlockTimestamp(pub u64);
-impl From<BlockTimestamp> for starknet::BlockTimestamp {
-    fn from(val: BlockTimestamp) -> Self {
-        starknet::BlockTimestamp(val.0)
-    }
-}
-
-#[derive(
-    Debug, Copy, Clone, Default, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord,
-)]
-pub struct ContractAddress(pub StarkHash);
-impl From<ContractAddress> for starknet::ContractAddress {
-    fn from(val: ContractAddress) -> Self {
-        starknet::ContractAddress(val.0.into())
-    }
-}
-
 #[derive(Debug, Default, Deserialize, Serialize, Clone, PartialEq)]
 pub struct DeployedContract {
     pub address: ContractAddress,
     pub class_hash: ClassHash,
-}
-
-#[derive(
-    Debug, Copy, Clone, Default, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord,
-)]
-#[serde(
-    from = "PrefixedHexAsBytes<16_usize>",
-    into = "PrefixedHexAsBytes<16_usize>"
-)]
-pub struct GasPrice(pub u128);
-impl From<PrefixedHexAsBytes<16_usize>> for GasPrice {
-    fn from(val: PrefixedHexAsBytes<16_usize>) -> Self {
-        GasPrice(u128::from_be_bytes(val.0))
-    }
-}
-impl From<GasPrice> for PrefixedHexAsBytes<16_usize> {
-    fn from(val: GasPrice) -> Self {
-        HexAsBytes(val.0.to_be_bytes())
-    }
-}
-impl From<GasPrice> for starknet::GasPrice {
-    fn from(val: GasPrice) -> Self {
-        starknet::GasPrice(val.0)
-    }
-}
-
-#[derive(
-    Debug, Copy, Clone, Default, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord,
-)]
-#[serde(
-    from = "NonPrefixedHexAsBytes<32_usize>",
-    into = "NonPrefixedHexAsBytes<32_usize>"
-)]
-pub struct GlobalRoot(pub StarkHash);
-// We don't use the regular StarkHash deserialization since the Starknet sequencer returns the
-// global root hash as a hex string without a "0x" prefix.
-impl From<NonPrefixedHexAsBytes<32_usize>> for GlobalRoot {
-    fn from(val: NonPrefixedHexAsBytes<32_usize>) -> Self {
-        GlobalRoot(StarkHash(val.0))
-    }
-}
-impl From<GlobalRoot> for NonPrefixedHexAsBytes<32_usize> {
-    fn from(val: GlobalRoot) -> Self {
-        HexAsBytes(val.0 .0)
-    }
-}
-impl From<GlobalRoot> for starknet::GlobalRoot {
-    fn from(val: GlobalRoot) -> Self {
-        starknet::GlobalRoot(val.0.into())
-    }
 }
 
 #[derive(Debug, Default, Deserialize, Serialize, Clone, PartialEq)]
