+++ src/starknet_client/objects/mod.rs
@@ -1,41 +1,2 @@
 pub mod block;
 pub mod transaction;
-
-use serde::{Deserialize, Serialize};
-
-use crate::starknet;
-use crate::starknet::serde_utils::{HexAsBytes, PrefixedHexAsBytes};
-
-#[derive(
-    Debug, Copy, Clone, Default, PartialEq, Eq, Hash, Deserialize, Serialize, PartialOrd, Ord,
-)]
-pub struct ContractAddress(pub StarkHash);
-impl From<ContractAddress> for starknet::ContractAddress {
-    fn from(val: ContractAddress) -> Self {
-        starknet::ContractAddress(val.0.into())
-    }
-}
-
-#[derive(
-    Debug, Copy, Clone, PartialEq, Eq, Default, Hash, Deserialize, Serialize, PartialOrd, Ord,
-)]
-#[serde(
-    from = "PrefixedHexAsBytes<32_usize>",
-    into = "PrefixedHexAsBytes<32_usize>"
-)]
-pub struct StarkHash(pub [u8; 32]);
-impl From<PrefixedHexAsBytes<32_usize>> for StarkHash {
-    fn from(val: PrefixedHexAsBytes<32_usize>) -> Self {
-        StarkHash(val.0)
-    }
-}
-impl From<StarkHash> for PrefixedHexAsBytes<32_usize> {
-    fn from(val: StarkHash) -> Self {
-        HexAsBytes(val.0)
-    }
-}
-impl From<StarkHash> for starknet::StarkHash {
-    fn from(val: StarkHash) -> Self {
-        starknet::StarkHash(val.0)
-    }
-}
